feat(ui): move server logs enhancements to info-view and add floating scroll buttons
