+++ packages/ui/src/components/info-view.tsx
@@ -1,6 +1,6 @@
 import { Component, For, createSignal, createEffect, Show, onMount, onCleanup, createMemo } from "solid-js"
-import { getInstanceLogs, instances, isInstanceLogStreaming, setInstanceLogStreaming } from "../stores/instances"
-import { ChevronDown } from "lucide-solid"
+import { getInstanceLogs, instances, isInstanceLogStreaming, setInstanceLogStreaming, clearLogs } from "../stores/instances"
+import { ChevronDown, Trash2 } from "lucide-solid"
 import InstanceInfo from "./instance-info"
 import { useI18n } from "../lib/i18n"
 
@@ -15,6 +15,8 @@ const InfoView: Component<InfoViewProps> = (props) => {
   let scrollRef: HTMLDivElement | undefined
   const savedState = logsScrollState.get(props.instanceId)
   const [autoScroll, setAutoScroll] = createSignal(savedState?.autoScroll ?? false)
+  const [showScrollTopButton, setShowScrollTopButton] = createSignal(false)
+  const [showScrollBottomButton, setShowScrollBottomButton] = createSignal(false)
 
   const instance = () => instances().get(props.instanceId)
   const logs = createMemo(() => getInstanceLogs(props.instanceId))
@@ -22,12 +24,15 @@ const InfoView: Component<InfoViewProps> = (props) => {
 
   const handleEnableLogs = () => setInstanceLogStreaming(props.instanceId, true)
   const handleDisableLogs = () => setInstanceLogStreaming(props.instanceId, false)
+  const handleClearLogs = () => clearLogs(props.instanceId)
  
   onMount(() => {
 
     if (scrollRef && savedState) {
       scrollRef.scrollTop = savedState.scrollTop
     }
+    // 初始化滾動按鈕可見性 / Initialize scroll button visibility
+    updateScrollButtons()
   })
 
   onCleanup(() => {
@@ -45,12 +50,37 @@ const InfoView: Component<InfoViewProps> = (props) => {
     }
   })
 
+  /** 更新滾動按鈕顯示狀態 / Update scroll button visibility */
+  const updateScrollButtons = () => {
+    if (!scrollRef) return
+
+    const scrollTop = scrollRef.scrollTop
+    const scrollHeight = scrollRef.scrollHeight
+    const clientHeight = scrollRef.clientHeight
+    const hasItems = logs().length > 0
+
+    const atBottom = scrollHeight - (scrollTop + clientHeight) <= 50
+    const atTop = scrollTop <= 50
+
+    setShowScrollBottomButton(hasItems && !atBottom)
+    setShowScrollTopButton(hasItems && !atTop)
+  }
+
+  /** 滾動至頂部 / Scroll to top */
+  const scrollToTop = () => {
+    if (scrollRef) {
+      scrollRef.scrollTop = 0
+      setAutoScroll(false)
+    }
+  }
+
   const handleScroll = () => {
     if (!scrollRef) return
 
     const isAtBottom = scrollRef.scrollHeight - scrollRef.scrollTop <= scrollRef.clientHeight + 50
 
     setAutoScroll(isAtBottom)
+    updateScrollButtons()
   }
 
   const scrollToBottom = () => {
@@ -83,6 +113,11 @@ const InfoView: Component<InfoViewProps> = (props) => {
     }
   }
 
+  /** 是否顯示浮動滾動按鈕 / Whether to show floating scroll buttons */
+  const showScrollButtons = createMemo(() => {
+    return streamingEnabled() && (showScrollTopButton() || showScrollBottomButton())
+  })
+
   return (
     <div class="log-container">
       <div class="flex-1 flex flex-col lg:flex-row gap-4 p-4 overflow-hidden">
@@ -94,6 +129,16 @@ const InfoView: Component<InfoViewProps> = (props) => {
           <div class="log-header">
             <h2 class="panel-title">{t("infoView.logs.title")}</h2>
             <div class="flex items-center gap-2">
+              <Show when={logs().length > 0}>
+                <button
+                  type="button"
+                  class="button-tertiary"
+                  onClick={handleClearLogs}
+                  title={t("infoView.logs.actions.clear")}
+                >
+                  <Trash2 class="w-4 h-4" />
+                </button>
+              </Show>
               <Show
                 when={streamingEnabled()}
                 fallback={
@@ -144,7 +189,7 @@ const InfoView: Component<InfoViewProps> = (props) => {
             </Show>
           </div>
  
-          <Show when={!autoScroll() && streamingEnabled()}>
+          <Show when={!autoScroll() && streamingEnabled() && logs().length > 0}>
             <button
               onClick={scrollToBottom}
               class="scroll-to-bottom"
@@ -152,6 +197,38 @@ const InfoView: Component<InfoViewProps> = (props) => {
               <ChevronDown class="w-4 h-4" />
               {t("infoView.logs.scrollToBottom")}
             </button>
+          </Show>
+
+          {/* 浮動滾動按鈕 / Floating scroll buttons */}
+          <Show when={showScrollButtons()}>
+            <div class="message-scroll-button-wrapper">
+              <Show when={showScrollTopButton()}>
+                <button
+                  type="button"
+                  class="message-scroll-button"
+                  onClick={scrollToTop}
+                  aria-label={t("infoView.logs.scrollToTop")}
+                  title={t("infoView.logs.scrollToTop")}
+                >
+                  <span class="message-scroll-icon" aria-hidden="true">
+                    ↑
+                  </span>
+                </button>
+              </Show>
+              <Show when={showScrollBottomButton()}>
+                <button
+                  type="button"
+                  class="message-scroll-button"
+                  onClick={scrollToBottom}
+                  aria-label={t("infoView.logs.scrollToBottom")}
+                  title={t("infoView.logs.scrollToBottom")}
+                >
+                  <span class="message-scroll-icon" aria-hidden="true">
+                    ↓
+                  </span>
+                </button>
+              </Show>
+            </div>
           </Show>
         </div>
       </div>
