+++ packages/ui/src/lib/i18n/messages/ja/logs.ts
@@ -14,9 +14,11 @@ export const logMessages = {
   "infoView.logs.title": "サーバーログ",
   "infoView.logs.actions.show": "サーバーログを表示",
   "infoView.logs.actions.hide": "サーバーログを非表示",
+  "infoView.logs.actions.clear": "ログをクリア",
   "infoView.logs.paused.title": "サーバーログは一時停止中です",
   "infoView.logs.paused.description": "ストリーミングを有効にして OpenCode サーバーの動作を監視します。",
   "infoView.logs.empty.waiting": "サーバー出力を待機中...",
+  "infoView.logs.scrollToTop": "最上部へスクロール",
   "infoView.logs.scrollToBottom": "最下部へスクロール",
 
   "infoView.dispose.actions.dispose": "インスタンスを破棄",
